feat:[007_4-Board-Attachment Strategy]
- 게시글 삭제시 첨부되어 있는 이미지 파일도 같이 삭제되도록 수정.
- 게시글 삭제 테스트 코드 수정

+++ src/main/java/com/project/usedItemsTrade/board/service/impl/BoardServiceImpl.java
@@ -142,15 +142,22 @@ public class BoardServiceImpl implements BoardService {
 
     @Override
     @Transactional
-    public void deleteBoard(Long id, String email) {
-        Board board = boardRepository.findById(id)
+    public void deleteBoard(BoardRequestDto.BoardDeleteDto deleteDto, String email) {
+        Board board = boardRepository.findById(deleteDto.getBoardId())
                 .orElseThrow(NoBoardExistsException::new);
 
         if (!board.getMember().getEmail().equals(email)) {
             throw new UserNotMatchException();
         }
 
-        imageRepository.deleteByBoard(board);
+        // 업로드된 이미지 삭제
+        if (deleteDto.getDeleteImageList() != null && deleteDto.getDeleteImageList().size() > 0) {
+            for (ImageDto.UploadResultDto resultDto : deleteDto.getDeleteImageList()) {
+                imageService.removeFile(resultDto.getImageURL());
+            }
+            imageRepository.deleteByBoard(board);
+        }
+
         boardRepository.delete(board);
     }
